% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise2.R
\name{mylatin}
\alias{mylatin}
\title{My own version of Pig Latin}
\usage{
mylatin(x)
}
\arguments{
\item{x}{An input string}
}
\value{
My version of the Pig Latin based on the input string
}
\description{
This function takes an input string then converts to my version of Pig Latin.
}
\examples{
mylatin("nba,nfl") 
mylatin("djfkdfjDFjR") 
}
